// finrisk-web/src/generated/api-schema.ts
export interface components {
    schemas: {
        /** @enum {string} */
        AssetType: "STOCK" | "ETF" | "BOND" | "CRYPTO";
// ...
        AssetResponse: components["schemas"]["StockResponse"] | components["schemas"]["EtfResponse"] | components["schemas"]["BondResponse"] | components["schemas"]["CryptoResponse"];
// ...
        AssetCreateRequest: components["schemas"]["StockCreateRequest"] | components["schemas"]["EtfCreateRequest"] | components["schemas"]["BondCreateRequest"] | components["schemas"]["CryptoCreateRequest"];
// ...
    };
// ...
}